Assignments/Assignment4/4requirements.txt

---

**`requirements.txt`**
```txt
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Optional: local document loaders
pypdf>=4.0.0
python-docx>=1.1.0
